Only write static files when they don't exist. waskiq/issue55

+++ ringo/config.py
@@ -224,16 +224,18 @@ def write_formbar_static_files():
     formbar_css = os.path.join(static_dir, 'formbar')
     for filename, content in get_formbar_css():
         filename = os.path.join(formbar_css, filename)
-        head, tail = os.path.split(filename)
-        if not os.path.exists(head):
-            os.makedirs(head)
-        with open(filename, 'wb') as f:
-            f.write(content)
+        if not os.path.isfile(filename):
+            head, tail = os.path.split(filename)
+            if not os.path.exists(head):
+                os.makedirs(head)
+            with open(filename, 'wb') as f:
+                f.write(content)
     formbar_js = os.path.join(static_dir, 'formbar')
     for filename, content in get_formbar_js():
         filename = os.path.join(formbar_js, filename)
-        head, tail = os.path.split(filename)
-        if not os.path.exists(head):
-            os.makedirs(head)
-        with open(filename, 'wb') as f:
-            f.write(content)
+        if not os.path.isfile(filename):
+            head, tail = os.path.split(filename)
+            if not os.path.exists(head):
+                os.makedirs(head)
+            with open(filename, 'wb') as f:
+                f.write(content)
